fixed clippy, added taplo, enabled rustfmt...
...updated packages, made the crate a lib too
