+++ src/fake_transaction_sink.rs
@@ -15,16 +15,10 @@ use std::{
 };
 use tracing::trace;
 
-#[derive(Clone)]
+#[derive(Default, Clone)]
 pub struct FakeTransactionSink {
 	txs: Arc<RwLock<HashSet<FakeHash>>>,
 	pub(crate) nonces: Arc<RwLock<HashMap<String, u128>>>,
-}
-
-impl FakeTransactionSink {
-	pub fn new() -> Self {
-		Self { txs: Default::default(), nonces: Default::default() }
-	}
 }
 
 #[async_trait]
@@ -79,7 +73,7 @@ mod test {
 
 	#[tokio::test]
 	async fn fake_sink_submit_and_watch_works() {
-		let rpc = FakeTransactionSink::new();
+		let rpc = FakeTransactionSink::default();
 		let t = FakeTransaction::new_finalizable(1);
 
 		let t: Box<dyn Transaction<HashType = FakeHash>> = Box::from(t);
@@ -91,7 +85,7 @@ mod test {
 		assert_eq!(
 			v,
 			vec![
-				TransactionStatus::Broadcasted(3),
+				TransactionStatus::Broadcasted,
 				TransactionStatus::Validated,
 				TransactionStatus::InBlock(1u32.to_le_bytes().into()),
 				TransactionStatus::InBlock(2u32.to_le_bytes().into()),
@@ -103,7 +97,7 @@ mod test {
 
 	#[tokio::test]
 	async fn fake_sink_submit_work_droppable() {
-		let rpc = FakeTransactionSink::new();
+		let rpc = FakeTransactionSink::default();
 		let t = FakeTransaction::new_droppable(1, 10);
 		let t: Box<dyn Transaction<HashType = FakeHash>> = Box::from(t);
 		let r = rpc.submit(&*t).await.unwrap_err();
@@ -112,7 +106,7 @@ mod test {
 
 	#[tokio::test]
 	async fn fake_sink_submit_work_invalid() {
-		let rpc = FakeTransactionSink::new();
+		let rpc = FakeTransactionSink::default();
 		let t = FakeTransaction::new_invalid(1, 10);
 		let t: Box<dyn Transaction<HashType = FakeHash>> = Box::from(t);
 		let r = rpc.submit(&*t).await.unwrap_err();
@@ -121,7 +115,7 @@ mod test {
 
 	#[tokio::test]
 	async fn fake_sink_submit_work_error() {
-		let rpc = FakeTransactionSink::new();
+		let rpc = FakeTransactionSink::default();
 		let t = FakeTransaction::new_error(1, 10);
 		let t: Box<dyn Transaction<HashType = FakeHash>> = Box::from(t);
 		let r = rpc.submit(&*t).await.unwrap_err();
@@ -130,7 +124,7 @@ mod test {
 
 	#[tokio::test]
 	async fn fake_sink_submit_work_valid() {
-		let rpc = FakeTransactionSink::new();
+		let rpc = FakeTransactionSink::default();
 		let t = FakeTransaction::new_finalizable_quick(111);
 		let t: Box<dyn Transaction<HashType = FakeHash>> = Box::from(t);
 		let r = rpc.submit(&*t).await.unwrap();
@@ -141,7 +135,7 @@ mod test {
 	async fn fake_sink_submit_work_valid_2() {
 		init_logger();
 		info!("start");
-		let rpc = Arc::from(FakeTransactionSink::new());
+		let rpc = Arc::from(FakeTransactionSink::default());
 		let t1 = FakeTransaction::new(111, vec![EventDef::finalized(11, 250)].into());
 		let t2 = FakeTransaction::new(222, vec![EventDef::finalized(12, 250)].into());
 		let t1: Box<dyn Transaction<HashType = FakeHash>> = Box::from(t1);
